feat: fix EN-pin resets under motor load, add tank-turn + pivot steering
Power fix:
ESP32 was resetting near full PWM despite an isolated 6V/12V dual-
battery setup. Traced via USB-power test (rules out supply sag
entirely) to motor noise coupling onto EN through a non-star ground
path — reset reason was POWERON_RESET, not brownout.
- Add 0.1-1uF cap on ESP32 EN->GND
- Add 0.1uF snubber caps across each motor's terminals
- Rebuild grounding as a single-point star (was partially daisy-chained)
- Simplify power arch: one 12V pack + 5A BEC, BEC feeds logic only
  (ESP32 + TB6612 VCC), VM comes straight off the 12V pack
Confirmed stable across full 0-255 PWM range.

Steering rework:
leftPWM/rightPWM were uint8_t (forward-only) with no way to represent
wheel direction, so steering could only taper the inner wheel down to
a floor (MIN_INNER_FACTOR) - never a real pivot, never a tight turn.
- Bump PACKET_VERSION 2 -> 3: leftPWM/rightPWM now int16_t
  (sign = direction, magnitude = duty)
- inputs.cpp: replace floor-based tapering with through-zero mixing
  (full fwd at center -> 0 at half lock -> full reverse at full lock)
- inputs.cpp: add idle-throttle + steer = in-place pivot, reusing
  existing pot/encoder, no new hardware
- outputs.cpp: derive TB6612 direction pins from PWM sign instead of
  hardcoding forward
Compiles clean; bench test pending (no hardware access yet).

Docs:
- README: rewrite Electrical Architecture for BEC-unified-logic
  topology, update wiring checklist, mark Competition Prep steering
  as code-complete pending bench test, flag packet.h v3 in repo notes
- devlog: add entries for EN-pin noise root-cause and packet v3
  signed-PWM rework

+++ src/bot_controller/transmitter/inputs.h
@@ -79,8 +79,8 @@
 // ---- Public data produced by this module -----------------------------
 struct InputState
 {
-    uint8_t  leftPWM;      // 0-255, final left motor duty (forward only)
-    uint8_t  rightPWM;     // 0-255, final right motor duty (forward only)
+    int16_t  leftPWM;      // -255..255, sign = direction, magnitude = duty
+    int16_t  rightPWM;     // -255..255, sign = direction, magnitude = duty
     uint8_t  speedPWM;     // 0-255, raw pot ceiling before steering is applied - for the display
     int8_t   steerSteps;   // clamped +/-STEER_MAX_STEPS - for the display
     uint8_t  buttons;      // bitmask, see packet.h BTN_* defines
